Revert "Move tests to DatagramUnicastInetTest because the JDK DatagramPacket does not support domain sockets"
This reverts commit a4432119e0f49ada792c6fe89bdc189501c04a51.
